requirements.txt
git+https://github.com/django/django.git
django-bootstrap3
django-allauth
git+https://github.com/django-debug-toolbar/django-debug-toolbar.git
python-memcached
https://bitbucket.org/russmo/django-storages-multi/get/default.zip
boto
django-ses
django-debug-toolbar
